deleted file _lof_0_22.py and _lof_1_0.py
